moved enterLocationActivity to order Activity package
