+++ Android/app/src/main/res/layout/activity_enter_location.xml
@@ -5,7 +5,7 @@
         xmlns:app="http://schemas.android.com/apk/res-auto"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".useractivities.EnterLocationActivity">
+        tools:context=".orderactivities.EnterLocationActivity">
 
     <EditText
             android:layout_width="wrap_content"
@@ -20,7 +20,7 @@
             android:layout_marginStart="99dp" android:autofillHints="Enter Table Number or Address"
             android:hint="@string/enter_table_number_or_address"/>
     <TextView
-            android:text=""
+            android:text="@string/please_enter_the_table_number_or_delivery_address_below"
             android:layout_width="200dp"
             android:layout_height="wrap_content" android:id="@+id/askEnterAddressMessage"
             app:layout_constraintTop_toTopOf="parent"
